fix(media): interpolate lock-screen progress and add GeckoView media session
Two #593 reports, one restructure.

Progress freeze: SystemUI interpolates the progress bar from the last
setState(position, speed) anchor, but pages managing position via
setPositionState() only report on play/pause/seek — never per second.
Meanwhile the 1s heartbeat kept re-publishing PlaybackState with the
stale positionSeconds, dragging SystemUI's anchor back every second:
the bar appeared frozen and jumped to the real position on pause. The
native session state machine now lives in an engine-agnostic
MediaSessionCore that anchors every position report with
elapsedRealtime, publishes position + elapsed × rate (clamped to
duration) while playing, freezes at the interpolated position when
leaving playback, and drops unchanged state reports instead of
re-anchoring. Non-playing publishes carry speed 0 per platform
convention.

GeckoView: the media bridge was android.webkit.WebView-only
(addJavascriptInterface + document-start injection), so Gecko builds
had no media session at all. GeckoView 142 implements the web Media
Session API itself, so a GeckoMediaSessionAdapter translates its event
delegate (onPlay/onPause/onStop, Metadata, PositionState, Feature
flags, Image.getBitmap artwork) into the same core and routes commands
back through the GeckoMediaSession controller — no JS polyfill, no
frame arbitration. GeckoViewEngine.setMediaSessionDelegate re-attaches
on every session rebuild; ShellActivity and WebViewActivity wire the
adapter on their Gecko branches behind the same enableMediaSession
flag. WebMediaPlaybackService now dispatches through the shared core,
so both engines ride one notification/service pipeline.

Fixes #593
