+++ app/src/main/java/com/webtoapp/core/webview/MediaSessionBridge.kt
@@ -1,74 +1,42 @@
 package com.webtoapp.core.webview
 
 import android.app.Activity
-import android.app.Notification
-import android.app.NotificationManager
-import android.app.PendingIntent
-import android.content.Context
-import android.content.Intent
-import android.graphics.Bitmap
-import android.graphics.BitmapFactory
-import android.media.MediaMetadata
-import android.media.session.MediaSession
 import android.media.session.PlaybackState
 import android.os.Handler
 import android.os.Looper
 import android.os.SystemClock
-import android.support.v4.media.session.MediaSessionCompat
 import android.webkit.JavascriptInterface
 import android.webkit.WebView
-import androidx.core.app.NotificationCompat
-import androidx.media.app.NotificationCompat.MediaStyle
-import com.webtoapp.core.logging.AppLogger
 import org.json.JSONObject
-import java.lang.ref.WeakReference
-import java.net.HttpURLConnection
-import java.net.URL
-import java.util.concurrent.Executors
-import kotlin.math.abs
 
 /**
  * Bridges the web [Media Session API](https://developer.mozilla.org/en-US/docs/Web/API/Media_Session_API)
- * (`navigator.mediaSession`) to a native Android [MediaSession].
+ * (`navigator.mediaSession`) of an Android [WebView] to a native media
+ * session owned by [MediaSessionCore].
  *
- * Audio is NOT transferred — it keeps playing inside the WebView. This bridge only
- * carries metadata, position and control commands between the page and the system
- * media controls (notification, lock screen, Bluetooth, Android Auto), backed by
- * [WebMediaPlaybackService] for background playback and wake lock.
+ * Audio is NOT transferred — it keeps playing inside the WebView. This bridge
+ * only carries metadata, position and control commands between the page and
+ * the system media controls (notification, lock screen, Bluetooth, Android
+ * Auto), backed by [WebMediaPlaybackService] for background playback and
+ * wake lock.
  *
  * The bridge constructor already registers the `WtaMediaSession` JavaScript
  * interface; callers only need to install [INJECTION_SCRIPT] at document start
- * (plus [injectNow] from `onPageFinished` as a fallback).
+ * (plus [injectNow] from `onPageFinished` as a fallback). Frame arbitration
+ * (#566) lives here; the native session state machine (position interpolation,
+ * teardown, notification) lives in the engine-agnostic core (#593).
  */
 class MediaSessionBridge(
-    private val activity: Activity,
+    activity: Activity,
     private val webView: WebView
 ) {
 
     private val mainHandler = Handler(Looper.getMainLooper())
-    private val artworkExecutor = Executors.newSingleThreadExecutor()
 
-    private val notificationManager =
-        activity.getSystemService(NotificationManager::class.java)
-
-    private val mediaSession = MediaSession(
+    private val core = MediaSessionCore(
         activity,
-        "WebToAppMediaSession"
+        ::dispatchCommand
     )
-
-    private var title = ""
-    private var artist = ""
-    private var album = ""
-    private var artworkUrl = ""
-
-    private var artworkBitmap: Bitmap? = null
-    private var loadedArtworkUrl = ""
-
-    private var durationSeconds = 0.0
-    private var positionSeconds = 0.0
-    private var playbackRate = 1.0f
-
-    private var playbackState = PlaybackState.STATE_NONE
 
     /**
      * Identity of the frame that currently owns the session. The injected script
@@ -81,75 +49,10 @@ class MediaSessionBridge(
     private var activeFrameId: String? = null
     private var activeFrameLastSeenMs = 0L
 
-    /**
-     * `true` once any non-`none` playback state has ever been reported. Used to
-     * guard [clearPlayback] so the initial `STATE_NONE` event (emitted during
-     * page load before any media exists) does not run full media-session
-     * teardown — which would otherwise stop the playback service on every cold
-     * start.
-     */
-    private var hasActivePlaybackSession = false
-
-    private var teardownRunnable: Runnable? = null
-
-    private val supportedActions = mutableSetOf<String>()
-
-    @Volatile
-    private var lastNotification: Notification? = null
-
     init {
-        activeBridge = WeakReference(this)
-
         webView.addJavascriptInterface(
             this,
             JAVASCRIPT_INTERFACE_NAME
-        )
-
-        createLaunchPendingIntent()?.let { pendingIntent ->
-            /*
-             * Opens or returns to the app when the user taps the
-             * notification, lock-screen media card, or system media player.
-             */
-            mediaSession.setSessionActivity(pendingIntent)
-        }
-
-        mediaSession.setCallback(
-            object : MediaSession.Callback() {
-                override fun onPlay() {
-                    dispatchCommand("play")
-                }
-
-                override fun onPause() {
-                    dispatchCommand("pause")
-                }
-
-                override fun onStop() {
-                    dispatchCommand("stop")
-                }
-
-                override fun onSeekTo(position: Long) {
-                    dispatchCommand(
-                        "seekto",
-                        position.toDouble() / 1000.0
-                    )
-                }
-
-                override fun onSkipToPrevious() {
-                    dispatchCommand("previoustrack")
-                }
-
-                override fun onSkipToNext() {
-                    dispatchCommand("nexttrack")
-                }
-
-                override fun onRewind() {
-                    dispatchCommand("seekbackward", 10.0)
-                }
-
-                override fun onFastForward() {
-                    dispatchCommand("seekforward", 10.0)
-                }
-            }
         )
     }
 
@@ -178,41 +81,7 @@ class MediaSessionBridge(
                 return@post
             }
 
-            val incomingBlank =
-                newTitle.isNullOrBlank() &&
-                    newArtist.isNullOrBlank() &&
-                    newAlbum.isNullOrBlank() &&
-                    newArtworkUrl.isNullOrBlank()
-
-            val currentBlank =
-                title.isBlank() &&
-                    artist.isBlank() &&
-                    album.isBlank() &&
-                    artworkUrl.isBlank()
-
-            /*
-             * A frame without media metadata used to send ("", "", "", "")
-             * every second, wiping real track info and falling back to the
-             * app label. Empty payloads never replace known metadata.
-             */
-            if (incomingBlank && !currentBlank) {
-                return@post
-            }
-
-            cancelTeardown()
-
-            title = newTitle.orEmpty()
-            artist = newArtist.orEmpty()
-            album = newAlbum.orEmpty()
-
-            val normalizedArtwork = newArtworkUrl.orEmpty()
-
-            if (artworkUrl != normalizedArtwork) {
-                artworkUrl = normalizedArtwork
-                loadArtwork(normalizedArtwork)
-            }
-
-            publish()
+            core.updateMetadata(newTitle, newArtist, newAlbum, newArtworkUrl)
         }
     }
 
@@ -249,13 +118,7 @@ class MediaSessionBridge(
                 return@post
             }
 
-            playbackState = resolved
-
-            if (resolved != PlaybackState.STATE_NONE) {
-                cancelTeardown()
-            }
-
-            publish()
+            core.updatePlaybackState(resolved)
         }
     }
 
@@ -275,32 +138,7 @@ class MediaSessionBridge(
                 activeFrameLastSeenMs = SystemClock.elapsedRealtime()
             }
 
-            val durationChanged =
-                abs(durationSeconds - duration) >= 0.5
-
-            durationSeconds =
-                duration.takeIf { it.isFinite() && it > 0.0 } ?: 0.0
-
-            positionSeconds =
-                position.takeIf { it.isFinite() && it >= 0.0 } ?: 0.0
-
-            playbackRate =
-                rate.takeIf { it.isFinite() && it > 0.0 }
-                    ?.toFloat()
-                    ?: 1.0f
-
-            /*
-             * Playback state must be updated regularly for progress tracking.
-             * Metadata is also republished when duration changes so the
-             * duration is never stuck at 0 when metadata arrived first.
-             */
-            applyPlaybackState()
-
-            if (durationChanged) {
-                applyMetadata()
-            }
-
-            updateNotification()
+            core.updatePosition(duration, position, rate)
         }
     }
 
@@ -309,26 +147,10 @@ class MediaSessionBridge(
         action: String?,
         supported: Boolean
     ) {
-        if (action.isNullOrBlank()) {
-            return
-        }
-
-        mainHandler.post {
-            if (supported) {
-                supportedActions.add(action)
-            } else {
-                supportedActions.remove(action)
-            }
-
-            /*
-             * Apply immediately instead of waiting for another playback event.
-             */
-            applyPlaybackState()
-            updateNotification()
-        }
+        core.setActionSupported(action, supported)
     }
 
-    // ---- State publishing ----
+    // ---- Frame arbitration ----
 
     /**
      * A frame owns its updates while it is the active frame and has reported
@@ -341,290 +163,11 @@ class MediaSessionBridge(
         return SystemClock.elapsedRealtime() - activeFrameLastSeenMs > FRAME_TAKEOVER_MS
     }
 
-    private fun cancelTeardown() {
-        teardownRunnable?.let { runnable ->
-            mainHandler.removeCallbacks(runnable)
-        }
-        teardownRunnable = null
-    }
-
-    private fun scheduleTeardown() {
-        cancelTeardown()
-
-        val runnable = Runnable {
-            teardownRunnable = null
-
-            if (playbackState == PlaybackState.STATE_NONE &&
-                (
-                    hasActivePlaybackSession ||
-                        mediaSession.isActive ||
-                        lastNotification != null
-                    )
-            ) {
-                clearPlayback()
-            }
-        }
-
-        teardownRunnable = runnable
-        mainHandler.postDelayed(runnable, TEARDOWN_GRACE_MS)
-    }
-
-    private fun publish() {
-        if (playbackState == PlaybackState.STATE_NONE) {
-            /*
-             * "none" arrives from any frame whose media ended (or that never
-             * had any). Tearing down immediately let one ended ad element kill
-             * a session another frame was still playing (#566); wait out the
-             * grace period and cancel if anything reports activity again.
-             */
-            if (hasActivePlaybackSession || mediaSession.isActive || lastNotification != null) {
-                scheduleTeardown()
-            }
-            return
-        }
-
-        cancelTeardown()
-
-        hasActivePlaybackSession = true
-
-        applyMetadata()
-        applyPlaybackState()
-
-        mediaSession.isActive = true
-
-        updateNotification()
-
-        WebMediaPlaybackService.synchronize(
-            activity.applicationContext,
-            playbackState == PlaybackState.STATE_PLAYING
-        )
-    }
-
-    private fun applyMetadata() {
-        val metadata = MediaMetadata.Builder()
-            .putString(
-                MediaMetadata.METADATA_KEY_TITLE,
-                title.ifBlank { appLabel() }
-            )
-            .putString(
-                MediaMetadata.METADATA_KEY_ARTIST,
-                artist
-            )
-            .putString(
-                MediaMetadata.METADATA_KEY_ALBUM,
-                album
-            )
-            .putLong(
-                MediaMetadata.METADATA_KEY_DURATION,
-                (durationSeconds * 1000.0).toLong()
-            )
-
-        artworkBitmap?.let { bitmap ->
-            metadata.putBitmap(
-                MediaMetadata.METADATA_KEY_ART,
-                bitmap
-            )
-
-            metadata.putBitmap(
-                MediaMetadata.METADATA_KEY_ALBUM_ART,
-                bitmap
-            )
-
-            metadata.putBitmap(
-                MediaMetadata.METADATA_KEY_DISPLAY_ICON,
-                bitmap
-            )
-        }
-
-        mediaSession.setMetadata(metadata.build())
-    }
-
-    private fun applyPlaybackState() {
-        var actions =
-            PlaybackState.ACTION_PLAY or
-                PlaybackState.ACTION_PAUSE or
-                PlaybackState.ACTION_PLAY_PAUSE or
-                PlaybackState.ACTION_STOP or
-                PlaybackState.ACTION_SEEK_TO
-
-        if ("previoustrack" in supportedActions) {
-            actions = actions or PlaybackState.ACTION_SKIP_TO_PREVIOUS
-        }
-
-        if ("nexttrack" in supportedActions) {
-            actions = actions or PlaybackState.ACTION_SKIP_TO_NEXT
-        }
-
-        if ("seekbackward" in supportedActions) {
-            actions = actions or PlaybackState.ACTION_REWIND
-        }
-
-        if ("seekforward" in supportedActions) {
-            actions = actions or PlaybackState.ACTION_FAST_FORWARD
-        }
-
-        val state = PlaybackState.Builder()
-            .setActions(actions)
-            .setState(
-                playbackState,
-                (positionSeconds * 1000.0).toLong(),
-                playbackRate
-            )
-            .build()
-
-        mediaSession.setPlaybackState(state)
-    }
-
-    // ---- Notification ----
-
-    private fun updateNotification() {
-        if (playbackState == PlaybackState.STATE_NONE) {
-            notificationManager.cancel(
-                WebMediaPlaybackService.NOTIFICATION_ID
-            )
-
-            lastNotification = null
-            return
-        }
-
-        val notification = buildNotification()
-
-        lastNotification = notification
-
-        notificationManager.notify(
-            WebMediaPlaybackService.NOTIFICATION_ID,
-            notification
-        )
-    }
-
-    private fun buildNotification(): Notification {
-        val builder = NotificationCompat.Builder(
-            activity,
-            WebMediaPlaybackService.CHANNEL_ID
-        )
-            .setSmallIcon(android.R.drawable.ic_media_play)
-            .setContentTitle(title.ifBlank { appLabel() })
-            .setContentText(
-                artist.ifBlank {
-                    album.ifBlank { "Media playback" }
-                }
-            )
-            .setSubText(album.takeIf { it.isNotBlank() })
-            .setCategory(NotificationCompat.CATEGORY_TRANSPORT)
-            .setVisibility(NotificationCompat.VISIBILITY_PUBLIC)
-            .setOnlyAlertOnce(true)
-            .setOngoing(
-                playbackState == PlaybackState.STATE_PLAYING
-            )
-            .setContentIntent(createLaunchPendingIntent())
-            .setDeleteIntent(
-                WebMediaPlaybackService.commandPendingIntent(
-                    activity,
-                    "stop"
-                )
-            )
-
-        artworkBitmap?.let(builder::setLargeIcon)
-
-        val compactActionIndexes = mutableListOf<Int>()
-        var actionIndex = 0
-
-        if ("previoustrack" in supportedActions) {
-            builder.addAction(
-                android.R.drawable.ic_media_previous,
-                "Previous",
-                WebMediaPlaybackService.commandPendingIntent(
-                    activity,
-                    "previoustrack"
-                )
-            )
-
-            compactActionIndexes.add(actionIndex)
-            actionIndex++
-        }
-
-        val isPlaying =
-            playbackState == PlaybackState.STATE_PLAYING
-
-        builder.addAction(
-            if (isPlaying) {
-                android.R.drawable.ic_media_pause
-            } else {
-                android.R.drawable.ic_media_play
-            },
-            if (isPlaying) "Pause" else "Play",
-            WebMediaPlaybackService.commandPendingIntent(
-                activity,
-                if (isPlaying) "pause" else "play"
-            )
-        )
-
-        compactActionIndexes.add(actionIndex)
-        actionIndex++
-
-        if ("nexttrack" in supportedActions) {
-            builder.addAction(
-                android.R.drawable.ic_media_next,
-                "Next",
-                WebMediaPlaybackService.commandPendingIntent(
-                    activity,
-                    "nexttrack"
-                )
-            )
-
-            compactActionIndexes.add(actionIndex)
-        }
-
-        val mediaStyle = MediaStyle()
-            .setMediaSession(
-                MediaSessionCompat.Token.fromToken(mediaSession.sessionToken)
-            )
-            .setShowCancelButton(true)
-            .setCancelButtonIntent(
-                WebMediaPlaybackService.commandPendingIntent(
-                    activity,
-                    "stop"
-                )
-            )
-
-        if (compactActionIndexes.isNotEmpty()) {
-            mediaStyle.setShowActionsInCompactView(
-                *compactActionIndexes
-                    .take(3)
-                    .toIntArray()
-            )
-        }
-
-        builder.setStyle(mediaStyle)
-
-        return builder.build()
-    }
-
-    private fun createLaunchPendingIntent(): PendingIntent? {
-        val launchIntent =
-            activity.packageManager.getLaunchIntentForPackage(
-                activity.packageName
-            ) ?: return null
-
-        launchIntent.addFlags(
-            Intent.FLAG_ACTIVITY_SINGLE_TOP or
-                Intent.FLAG_ACTIVITY_CLEAR_TOP
-        )
-
-        return PendingIntent.getActivity(
-            activity,
-            3101,
-            launchIntent,
-            PendingIntent.FLAG_UPDATE_CURRENT or
-                PendingIntent.FLAG_IMMUTABLE
-        )
-    }
-
     // ---- Android → JS commands ----
 
     private fun dispatchCommand(
         command: String,
-        value: Double = 0.0
+        value: Double
     ) {
         val escapedCommand = JSONObject.quote(command)
 
@@ -638,107 +181,8 @@ class MediaSessionBridge(
         }
     }
 
-    // ---- Artwork ----
-
-    private fun loadArtwork(url: String) {
-        if (url.isBlank()) {
-            artworkBitmap = null
-            loadedArtworkUrl = ""
-            publish()
-            return
-        }
-
-        if (url == loadedArtworkUrl && artworkBitmap != null) {
-            return
-        }
-
-        artworkExecutor.execute {
-            val bitmap = downloadBitmap(url)
-
-            mainHandler.post {
-                /*
-                 * Ignore an old request when the website changed tracks while
-                 * the image was still downloading.
-                 */
-                if (artworkUrl != url) {
-                    return@post
-                }
-
-                artworkBitmap = bitmap
-                loadedArtworkUrl = url
-
-                if (playbackState != PlaybackState.STATE_NONE) {
-                    applyMetadata()
-                    updateNotification()
-                }
-            }
-        }
-    }
-
-    private fun downloadBitmap(url: String): Bitmap? {
-        return try {
-            val connection = URL(url).openConnection()
-                as HttpURLConnection
-
-            connection.connectTimeout = 8_000
-            connection.readTimeout = 8_000
-            connection.instanceFollowRedirects = true
-            connection.doInput = true
-            connection.connect()
-
-            connection.inputStream.use {
-                BitmapFactory.decodeStream(it)
-            }.also {
-                connection.disconnect()
-            }
-        } catch (_: Exception) {
-            null
-        }
-    }
-
-    private fun appLabel(): String {
-        return activity.applicationInfo
-            .loadLabel(activity.packageManager)
-            .toString()
-    }
-
-    private fun clearPlayback() {
-        cancelTeardown()
-
-        hasActivePlaybackSession = false
-
-        activeFrameId = null
-        activeFrameLastSeenMs = 0L
-
-        playbackState = PlaybackState.STATE_NONE
-        positionSeconds = 0.0
-        durationSeconds = 0.0
-
-        mediaSession.setPlaybackState(
-            PlaybackState.Builder()
-                .setState(
-                    PlaybackState.STATE_NONE,
-                    0L,
-                    1.0f
-                )
-                .build()
-        )
-
-        mediaSession.isActive = false
-
-        notificationManager.cancel(
-            WebMediaPlaybackService.NOTIFICATION_ID
-        )
-
-        lastNotification = null
-
-        WebMediaPlaybackService.stop(
-            activity.applicationContext
-        )
-    }
-
     fun release() {
-        clearPlayback()
+        core.release()
 
         try {
             webView.removeJavascriptInterface(
@@ -746,13 +190,6 @@ class MediaSessionBridge(
             )
         } catch (_: Exception) {
             // WebView may already be destroyed.
-        }
-
-        artworkExecutor.shutdownNow()
-        mediaSession.release()
-
-        if (activeBridge?.get() === this) {
-            activeBridge = null
         }
     }
 
@@ -762,31 +199,9 @@ class MediaSessionBridge(
 
         /**
          * How long the active frame's reports stay authoritative after the
-         * last one, and how long a `none` report must survive before the
-         * session is torn down.
+         * last one.
          */
         private const val FRAME_TAKEOVER_MS = 3_000L
-        private const val TEARDOWN_GRACE_MS = 2_500L
-
-        @Volatile
-        private var activeBridge:
-            WeakReference<MediaSessionBridge>? = null
-
-        internal fun dispatchFromService(
-            command: String,
-            value: Double
-        ) {
-            activeBridge?.get()?.dispatchCommand(
-                command,
-                value
-            )
-        }
-
-        internal fun getForegroundNotification(
-            context: Context
-        ): Notification? {
-            return activeBridge?.get()?.lastNotification
-        }
 
         /** The document-start JS polyfill that hooks navigator.mediaSession. */
         fun getInjectionScript(): String = INJECTION_SCRIPT
@@ -1369,7 +784,9 @@ class MediaSessionBridge(
 
               /*
                * A one-second heartbeat keeps lock-screen progress accurate and
-               * detects metadata changes made by the website.
+               * detects metadata changes made by the website. The native core
+               * deduplicates unchanged state reports and interpolates position
+               * between them (#593).
                */
               window.__wtaMediaHeartbeat =
                 window.setInterval(
